+++ weblate/trans/models/component.py
@@ -2531,7 +2531,7 @@ class Component(models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin):
                 raise ValidationError(
                     {"category": gettext("Category does not belong to this project.")}
                 )
-            if self.links.exists():
+            if self.pk and self.links.exists():
                 message = gettext("Categorized component can not be shared.")
                 raise ValidationError({"category": message, "links": message})